Move common MP Config message to a common bundle
