--- dev/io.openliberty.io.smallrye.config/resources/io/openliberty/io/smallrye/config/resources/SmallryeConfigMessages_es.nlsprops
@@ -1,28 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.io.smallrye.config/resources/io/openliberty/io/smallrye/config/resources/SmallryeConfigMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-
-#COMPONENTPREFIX CWWKC                 
-#COMPONENTNAMEFOR CWWKC Smallrye Config
-
-# Message numbers CWWKC0651-700 are reserved for this file.
-# NLS_MESSAGEFORMAT_VAR                
-# MSG_DESCRIPTIVE_NAME_CWWKC0000=CWWKC0000I: This is a message with inserts {0}
-# MSG_DESCRIPTIVE_NAME_CWWKC0000.explanation=Explanation text for the message
-# MSG_DESCRIPTIVE_NAME_CWWKC0000.useraction=User action text for the message
-
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651=CWWKC0651W: El valor de configuraci\u00f3n MicroProfile para la clave {0} ha cambiado desde que se complet\u00f3 la acci\u00f3n de punto de comprobaci\u00f3n en el servidor. Si el valor de la clave cambia despu\u00e9s de la acci\u00f3n de punto de comprobaci\u00f3n, es posible que la aplicaci\u00f3n no utilice el valor actualizado.
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.explanation=Se ha accedido al valor de configuraci\u00f3n durante una acci\u00f3n de punto de comprobaci\u00f3n y se ha actualizado antes de que se restaurara el proceso del servidor. En funci\u00f3n de c\u00f3mo acceda la aplicaci\u00f3n a la configuraci\u00f3n, es posible que no se utilice el valor actualizado.
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.useraction=Evite acceder a las propiedades de configuraci\u00f3n durante un punto de comprobaci\u00f3n.
