--- dev/io.openliberty.io.smallrye.config/resources/io/openliberty/io/smallrye/config/resources/SmallryeConfigMessages_ro.nlsprops
@@ -1,28 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.io.smallrye.config/resources/io/openliberty/io/smallrye/config/resources/SmallryeConfigMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-
-#COMPONENTPREFIX CWWKC                 
-#COMPONENTNAMEFOR CWWKC Smallrye Config
-
-# Message numbers CWWKC0651-700 are reserved for this file.
-# NLS_MESSAGEFORMAT_VAR                
-# MSG_DESCRIPTIVE_NAME_CWWKC0000=CWWKC0000I: This is a message with inserts {0}
-# MSG_DESCRIPTIVE_NAME_CWWKC0000.explanation=Explanation text for the message
-# MSG_DESCRIPTIVE_NAME_CWWKC0000.useraction=User action text for the message
-
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651=CWWKC0651W: Valoarea de configura\u021bie MicroProfile pentru cheia {0} s-a modificat de la finalizarea ac\u021biunii de punct de control pentru server. \u00cen cazul \u00een care valoarea cheii se modific\u0103 dup\u0103 ac\u021biunea punctului de control, este posibil ca aplica\u021bia s\u0103 nu poat\u0103 utiliza valoarea actualizat\u0103.
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.explanation=Valoarea configura\u021biei a fost accesat\u0103 \u00een timpul unei ac\u021biuni a punctului de control \u0219i actualizat\u0103 \u00eenainte ca procesul de server s\u0103 fie restabilit. \u00cen func\u021bie de felul \u00een care aplica\u021bia acceseaz\u0103 configura\u021bia, s-ar putea ca valoarea actualizat\u0103 s\u0103 nu poat\u0103 fi utilizat\u0103.
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.useraction=Evita\u021bi accesarea propriet\u0103\u021bilor de configurare \u00een timpul punctului de control.
