--- dev/io.openliberty.io.smallrye.config3/resources/io/openliberty/io/smallrye/config/resources/SmallryeConfigMessages_cs.nlsprops
@@ -1,28 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.io.smallrye.config/resources/io/openliberty/io/smallrye/config/resources/SmallryeConfigMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-
-#COMPONENTPREFIX CWWKC                 
-#COMPONENTNAMEFOR CWWKC Smallrye Config
-
-# Message numbers CWWKC0651-700 are reserved for this file.
-# NLS_MESSAGEFORMAT_VAR                
-# MSG_DESCRIPTIVE_NAME_CWWKC0000=CWWKC0000I: This is a message with inserts {0}
-# MSG_DESCRIPTIVE_NAME_CWWKC0000.explanation=Explanation text for the message
-# MSG_DESCRIPTIVE_NAME_CWWKC0000.useraction=User action text for the message
-
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651=CWWKC0651W: Konfigura\u010dn\u00ed hodnota MicroProfile pro kl\u00ed\u010d {0} se zm\u011bnila od dokon\u010den\u00ed akce kontroln\u00edho bodu na serveru. Pokud se hodnota kl\u00ed\u010dov\u00fdch zm\u011bn po proveden\u00ed akce kontroln\u00edho bodu zm\u011bn\u00ed, aplikace nemus\u00ed pou\u017e\u00edt aktualizovanou hodnotu.
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.explanation=Ke konfigura\u010dn\u00ed hodnot\u011b bylo p\u0159istoupeno b\u011bhem akce kontroln\u00edho bodu a byla aktualizov\u00e1na, ne\u017e se obnovil proces serveru. V z\u00e1vislosti na tom, jak aplikace p\u0159istupuje ke konfiguraci, nemus\u00ed b\u00fdt aktualizovan\u00e1 hodnota pou\u017eita.
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.useraction=Vyvarujte se p\u0159\u00edstupu ke konfigura\u010dn\u00edm vlastnostem b\u011bhem kontroln\u00edho bodu.
