--- dev/io.openliberty.io.smallrye.config3/resources/io/openliberty/io/smallrye/config/resources/SmallryeConfigMessages_ko.nlsprops
@@ -1,28 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.io.smallrye.config/resources/io/openliberty/io/smallrye/config/resources/SmallryeConfigMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-
-#COMPONENTPREFIX CWWKC                 
-#COMPONENTNAMEFOR CWWKC Smallrye Config
-
-# Message numbers CWWKC0651-700 are reserved for this file.
-# NLS_MESSAGEFORMAT_VAR                
-# MSG_DESCRIPTIVE_NAME_CWWKC0000=CWWKC0000I: This is a message with inserts {0}
-# MSG_DESCRIPTIVE_NAME_CWWKC0000.explanation=Explanation text for the message
-# MSG_DESCRIPTIVE_NAME_CWWKC0000.useraction=User action text for the message
-
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651=CWWKC0651W: \uc11c\ubc84\uc5d0\uc11c \uccb4\ud06c\ud3ec\uc778\ud2b8 \uc870\uce58\uac00 \uc644\ub8cc\ub41c \uc774\ud6c4 {0} \ud0a4\uc5d0 \ub300\ud55c MicroProfile \uad6c\uc131 \uac12\uc774 \ubcc0\uacbd\ub418\uc5c8\uc2b5\ub2c8\ub2e4. \uccb4\ud06c\ud3ec\uc778\ud2b8 \uc870\uce58 \ud6c4 \ud0a4 \uac12\uc774 \ubcc0\uacbd\ub418\uba74 \uc560\ud50c\ub9ac\ucf00\uc774\uc158\uc774 \uc5c5\ub370\uc774\ud2b8\ub41c \uac12\uc744 \uc0ac\uc6a9\ud558\uc9c0 \uc54a\uc744 \uc218 \uc788\uc2b5\ub2c8\ub2e4.
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.explanation=\uad6c\uc131 \uac12\uc740 \uccb4\ud06c\ud3ec\uc778\ud2b8 \uc870\uce58 \uc911\uc5d0 \uc561\uc138\uc2a4\ub418\uc5c8\uc73c\uba70 \uc11c\ubc84 \ud504\ub85c\uc138\uc2a4\uac00 \ubcf5\uc6d0\ub418\uae30 \uc804\uc5d0 \uc5c5\ub370\uc774\ud2b8\ub418\uc5c8\uc2b5\ub2c8\ub2e4. \uc560\ud50c\ub9ac\ucf00\uc774\uc158\uc774 \uad6c\uc131\uc5d0 \uc561\uc138\uc2a4\ud558\ub294 \ubc29\ubc95\uc5d0 \ub530\ub77c \uc5c5\ub370\uc774\ud2b8\ub41c \uac12\uc774 \uc0ac\uc6a9\ub418\uc9c0 \uc54a\uc744 \uc218 \uc788\uc2b5\ub2c8\ub2e4.
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.useraction=\uccb4\ud06c\ud3ec\uc778\ud2b8 \uc911\uc5d0 \uad6c\uc131 \ud2b9\uc131\uc5d0 \uc561\uc138\uc2a4\ud558\uc9c0 \ub9c8\uc2ed\uc2dc\uc624.
