--- dev/io.openliberty.io.smallrye.config3/resources/io/openliberty/io/smallrye/config/resources/SmallryeConfigMessages_zh_TW.nlsprops
@@ -1,28 +1,0 @@
-###############################################################################
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License 2.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-2.0/
-# 
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-###############################################################################
-#CMVCPATHNAME io.openliberty.io.smallrye.config/resources/io/openliberty/io/smallrye/config/resources/SmallryeConfigMessages.nlsprops
-#ISMESSAGEFILE TRUE
-#NLS_ENCODING=UNICODE
-
-#COMPONENTPREFIX CWWKC                 
-#COMPONENTNAMEFOR CWWKC Smallrye Config
-
-# Message numbers CWWKC0651-700 are reserved for this file.
-# NLS_MESSAGEFORMAT_VAR                
-# MSG_DESCRIPTIVE_NAME_CWWKC0000=CWWKC0000I: This is a message with inserts {0}
-# MSG_DESCRIPTIVE_NAME_CWWKC0000.explanation=Explanation text for the message
-# MSG_DESCRIPTIVE_NAME_CWWKC0000.useraction=User action text for the message
-
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651=CWWKC0651W: \u81ea\u6aa2\u67e5\u9ede\u52d5\u4f5c\u5728\u4f3a\u670d\u5668\u4e0a\u5b8c\u6210\u4ee5\u4f86\uff0c{0} \u7d22\u5f15\u9375\u7684 MicroProfile \u914d\u7f6e\u503c\u5df2\u8b8a\u66f4\u3002 \u5982\u679c\u5728\u6aa2\u67e5\u9ede\u52d5\u4f5c\u4e4b\u5f8c\u8b8a\u66f4\u91d1\u9470\u7684\u503c\uff0c\u5247\u61c9\u7528\u7a0b\u5f0f\u53ef\u80fd\u4e0d\u6703\u4f7f\u7528\u66f4\u65b0\u7684\u503c\u3002
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.explanation=\u5728\u6aa2\u67e5\u9ede\u52d5\u4f5c\u671f\u9593\u5b58\u53d6\u914d\u7f6e\u503c\uff0c\u4e26\u5728\u9084\u539f\u4f3a\u670d\u5668\u8655\u7406\u7a0b\u5e8f\u4e4b\u524d\u66f4\u65b0\u914d\u7f6e\u503c\u3002 \u8996\u61c9\u7528\u7a0b\u5f0f\u5b58\u53d6\u914d\u7f6e\u7684\u65b9\u5f0f\u800c\u5b9a\uff0c\u53ef\u80fd\u4e0d\u6703\u4f7f\u7528\u66f4\u65b0\u7684\u503c\u3002
-WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.useraction=\u907f\u514d\u5728\u6aa2\u67e5\u9ede\u671f\u9593\u5b58\u53d6\u914d\u7f6e\u5167\u5bb9\u3002
