+++ dev/io.openliberty.microprofile.config.internal.common/bnd.bnd
@@ -1,14 +1,11 @@
 #*******************************************************************************
-# Copyright (c) 2020, 2021 IBM Corporation and others.
+# Copyright (c) 2020, 2023 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
 # http://www.eclipse.org/legal/epl-2.0/
 # 
 # SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
 #*******************************************************************************
 -include= ~../cnf/resources/bnd/bundle.props
 bVersion=1.0
@@ -19,7 +16,8 @@ Bundle-SymbolicName: io.openliberty.microprofile.config.internal.common
 
 Bundle-Description:MicroProfile Configuration common Implementation, version ${bVersion}
 
-Export-Package: io.openliberty.microprofile.config.internal.common
+Export-Package: io.openliberty.microprofile.config.internal.common,\
+                io.openliberty.microprofile.config.internal.common.resources
 
 Import-Package: \
         org.eclipse.microprofile.config;version="[1.0,4)", \
