+++ dev/io.openliberty.microprofile.config.internal.common/resources/io/openliberty/microprofile/config/internal/common/resources/MPConfigCommonMessages_de.nlsprops
@@ -1,0 +1,29 @@
+###############################################################################
+# Copyright (c) 2022, 2023 IBM Corporation and others.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License 2.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-2.0/
+# 
+# SPDX-License-Identifier: EPL-2.0
+###############################################################################
+#CMVCPATHNAME io.openliberty.microprofile.config.internal.common/resources/io/openliberty/microprofile/config/internal/common/resources/MPConfigCommonMessages.nlsprops
+#ISMESSAGEFILE TRUE
+#NLS_ENCODING=UNICODE
+
+#COMPONENTPREFIX CWWKC                 
+#COMPONENTNAMEFOR CWWKC MicroProfile Config with InstantOn
+# Message numbers CWWKC0651-700 are reserved for this file.
+
+#COMPONENTPREFIX CWMCG                 
+#COMPONENTNAMEFOR CWMCG MicroProfile Config Common
+# Message numbers CWMCG0400-499 are reserved for this file.
+
+# NLS_MESSAGEFORMAT_VAR                
+# MSG_DESCRIPTIVE_NAME_CWMCG0000=CWMCG0000I: This is a message with inserts {0}
+# MSG_DESCRIPTIVE_NAME_CWMCG0000.explanation=Explanation text for the message
+# MSG_DESCRIPTIVE_NAME_CWMCG0000.useraction=User action text for the message
+
+WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651=CWWKC0651W: Der MicroProfile-Konfigurationswert f\u00fcr den Schl\u00fcssel {0} hat sich seit Abschluss der Pr\u00fcfpunktaktion auf dem Server ge\u00e4ndert. Wenn sich der Wert des Schl\u00fcssels nach der Pr\u00fcfpunktaktion \u00e4ndert, verwendet die Anwendung m\u00f6glicherweise nicht den aktualisierten Wert.
+WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.explanation=Auf den Konfigurationswert wurde w\u00e4hrend einer Pr\u00fcfpunktaktion zugegriffen und er wurde aktualisiert, bevor der Serverprozess wiederhergestellt wurde. Je nachdem, wie die Anwendung auf die Konfiguration zugreift, wird der aktualisierte Wert m\u00f6glicherweise nicht verwendet.
+WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.useraction=Vermeiden Sie den Zugriff auf Konfigurationseigenschaften w\u00e4hrend des Pr\u00fcfpunkts.
