+++ dev/io.openliberty.microprofile.config.internal.common/resources/io/openliberty/microprofile/config/internal/common/resources/MPConfigCommonMessages_zh.nlsprops
@@ -1,0 +1,29 @@
+###############################################################################
+# Copyright (c) 2022, 2023 IBM Corporation and others.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License 2.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-2.0/
+# 
+# SPDX-License-Identifier: EPL-2.0
+###############################################################################
+#CMVCPATHNAME io.openliberty.microprofile.config.internal.common/resources/io/openliberty/microprofile/config/internal/common/resources/MPConfigCommonMessages.nlsprops
+#ISMESSAGEFILE TRUE
+#NLS_ENCODING=UNICODE
+
+#COMPONENTPREFIX CWWKC                 
+#COMPONENTNAMEFOR CWWKC MicroProfile Config with InstantOn
+# Message numbers CWWKC0651-700 are reserved for this file.
+
+#COMPONENTPREFIX CWMCG                 
+#COMPONENTNAMEFOR CWMCG MicroProfile Config Common
+# Message numbers CWMCG0400-499 are reserved for this file.
+
+# NLS_MESSAGEFORMAT_VAR                
+# MSG_DESCRIPTIVE_NAME_CWMCG0000=CWMCG0000I: This is a message with inserts {0}
+# MSG_DESCRIPTIVE_NAME_CWMCG0000.explanation=Explanation text for the message
+# MSG_DESCRIPTIVE_NAME_CWMCG0000.useraction=User action text for the message
+
+WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651=CWWKC0651W: \u81ea\u5728\u670d\u52a1\u5668\u4e0a\u5b8c\u6210\u68c0\u67e5\u70b9\u64cd\u4f5c\u4ee5\u6765\uff0c{0} \u5bc6\u94a5\u7684 MicroProfile \u914d\u7f6e\u503c\u5df2\u66f4\u6539\u3002 \u5982\u679c\u5728\u6267\u884c\u68c0\u67e5\u70b9\u64cd\u4f5c\u540e\u8be5\u952e\u7684\u503c\u53d1\u751f\u66f4\u6539\uff0c\u90a3\u4e48\u5e94\u7528\u7a0b\u5e8f\u53ef\u80fd\u672a\u4f7f\u7528\u66f4\u65b0\u540e\u7684\u503c\u3002
+WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.explanation=\u5728\u6267\u884c\u68c0\u67e5\u70b9\u64cd\u4f5c\u671f\u95f4\u8bbf\u95ee\u4e86\u8be5\u914d\u7f6e\u503c\uff0c\u5e76\u5728\u590d\u539f\u670d\u52a1\u5668\u8fdb\u7a0b\u4e4b\u524d\u66f4\u65b0\u4e86\u8be5\u914d\u7f6e\u503c\u3002 \u6839\u636e\u5e94\u7528\u7a0b\u5e8f\u8bbf\u95ee\u914d\u7f6e\u7684\u65b9\u5f0f\uff0c\u53ef\u80fd\u4e0d\u4f1a\u4f7f\u7528\u66f4\u65b0\u540e\u7684\u503c\u3002
+WARNING_UPDATED_CONFIG_PROPERTY_NOT_USED_CWWKC0651.useraction=\u8bf7\u907f\u514d\u5728\u6267\u884c\u68c0\u67e5\u70b9\u64cd\u4f5c\u671f\u95f4\u8bbf\u95ee\u914d\u7f6e\u5c5e\u6027\u3002
