+++ dev/io.openliberty.microprofile.config.internal.common/src/io/openliberty/microprofile/config/internal/common/resources/package-info.java
@@ -1,14 +1,11 @@
 /*******************************************************************************
- * Copyright (c) 2020, 2023 IBM Corporation and others.
+ * Copyright (c) 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
  *
  * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- * IBM Corporation - initial API and implementation
  *******************************************************************************/
 
 /**
@@ -16,6 +13,6 @@
  */
 @org.osgi.annotation.versioning.Version("1.0")
 @TraceOptions(traceGroup = "APPCONFIG", messageBundle = "io.openliberty.microprofile.config.internal.common.resources")
-package io.openliberty.microprofile.config.internal.common;
+package io.openliberty.microprofile.config.internal.common.resources;
 
 import com.ibm.websphere.ras.annotation.TraceOptions;
